requirements.txt
tk
Pillow
opencv-python
pytesseract
deep-translator
langdetect
numpy
fpdf
python-docx
